fixed IOException
git-svn-id: https://sovereigncomputing.net/svn/sneer/trunk@1677 ee3b4d21-14f5-0310-a1fa-c5d8b2925ec4

+++ kernel/src/sneer/lego/utils/FileUtils.java
@@ -33,6 +33,9 @@ public class FileUtils {
 	}
 	
 	public static void cleanDirectory(File directory) {
+		if(!directory.exists())
+			return;
+		
 		try {
 			org.apache.commons.io.FileUtils.cleanDirectory(directory);
 		} catch (IOException e) {
